backend/requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements/base.in
#
asgiref==3.11.0
    # via
    #   django
    #   django-allauth
    #   django-cors-headers
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
dj-database-url==3.1.0
    # via environs
dj-email-url==1.0.6
    # via environs
dj-rest-auth==7.0.2
    # via -r requirements/base.in
django==5.2.10
    # via
    #   -r requirements/base.in
    #   dj-database-url
    #   dj-rest-auth
    #   django-allauth
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-allauth==65.14.0
    # via -r requirements/base.in
django-cache-url==3.4.6
    # via environs
django-cors-headers==4.9.0
    # via -r requirements/base.in
djangorestframework==3.16.1
    # via
    #   -r requirements/base.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.1
    # via -r requirements/base.in
environs[django]==14.5.0
    # via -r requirements/base.in
gunicorn==24.1.1
    # via -r requirements/base.in
idna==3.11
    # via requests
marshmallow==4.2.1
    # via environs
packaging==26.0
    # via gunicorn
psycopg[binary]==3.3.2
    # via -r requirements/base.in
psycopg-binary==3.3.2
    # via psycopg
pyjwt==2.11.0
    # via djangorestframework-simplejwt
python-dotenv==1.2.1
    # via environs
requests==2.32.5
    # via -r requirements/base.in
sqlparse==0.5.5
    # via django
typing-extensions==4.15.0
    # via psycopg
urllib3==2.6.3
    # via requests
whitenoise==6.11.0
    # via -r requirements/base.in
